added

+++ Backend/apisentinel_net/DTOs/PessoaDTO.cs
@@ -1,9 +1,11 @@
+using System.Numerics;
+
 namespace DTOs;
 
 public class PessoaDTO
 {
     public int Id { get; set; }
-    public int CPF { get; set; }
+    public long CPF { get; set; }
     public int pagamento { get; set;  }
     
 }
